+++ plugins/code-enhancers-and-utils/kotlin-config/main/kotlin/KotlinConfigPlugin.kt
@@ -1,0 +1,11 @@
+package com.javiersc.gradle.plugins.kotlin.config
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class KotlinConfigPlugin : Plugin<Project> {
+
+    override fun apply(project: Project) {
+        KotlinConfigType.build(project)
+    }
+}