+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Updated
 
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.25`
+
 ## [0.1.0-rc.25] - 2021-12-24
 
 ### Added